adding jwt

+++ Models/DTO/ProductDTO.cs
@@ -6,6 +6,6 @@
         public required string Name { get; set; }
         public int Quntity { get; set; }
         public string Description { get; set; } = string.Empty;
-        public double Price { get; set; } = 0.00;
+        public decimal Price { get; set; } = 0;
     }
 }
